Refactor import paths for evaluation utilities; add quiz count query to QuizDao
